src/test/java/shamrokh/muhammad/leetcode/solution/SolutionTest.java
package shamrokh.muhammad.leetcode.solution;

import org.junit.jupiter.api.*;


import static org.junit.jupiter.api.Assertions.*;


@TestMethodOrder(MethodOrderer.OrderAnnotation.class)
class SolutionTest {
    Solution solution;
    @BeforeEach
    void setup(){
        solution = new Solution();
    }

    @Test
    @Order(1)
    @DisplayName("Test 1")
    void TestOne(){
        //Input: s = "leEeetcode"
        //Output: "leetcode"
        assertEquals("leetcode", solution.makeGood("leEeetcode"));
    }

    @Test
    @Order(2)
    @DisplayName("Test 2")
    void TestTwo(){
        //Input: s = "abBAcC"
        //Output: ""
        assertEquals("", solution.makeGood("abBAcC"));
    }

    @Test
    @Order(3)
    @DisplayName("Test 3")
    void TestThree(){
        //Input: s = "s"
        //Output: "s"
        assertEquals("s", solution.makeGood("s"));
    }

    @Test
    @Order(4)
    @DisplayName("Test 4")
    void TestFour(){
    }

}

